+++ flutter_inappwebview/example/integration_test/in_app_webview/main.dart
@@ -55,6 +55,7 @@ part 'javascript_handler_reply.dart';
 part 'load_data.dart';
 part 'load_file.dart';
 part 'load_file_url.dart';
+part 'load_file_url_request.dart';
 part 'load_url.dart';
 part 'on_console_message.dart';
 part 'on_content_size_changed.dart';
@@ -130,6 +131,7 @@ void main() {
     javascriptCodeEvaluation();
     loadUrl();
     loadFileUrl();
+    loadFileUrlRequest();
     javascriptHandler();
     javascriptHandlerError();
     javascriptHandlerReplyTypes();